Move NodeSelection to its own file
Classes defined before forms cause resource loading errors
